+++ packages/leancode_contracts/test/types/time_only_test.dart
@@ -8,19 +8,30 @@ void main() {
   group('TimeOnly', () {
     test('accepts correct durations', () {
       expect(
-        TimeOnly(const Duration(days: 1) - const Duration(microseconds: 1)),
+        TimeOnly.fromDuration(
+          const Duration(days: 1) - const Duration(microseconds: 1),
+        ),
         isA<TimeOnly>(),
       );
-      expect(TimeOnly(Duration.zero), isA<TimeOnly>());
-      expect(TimeOnly(const Duration(hours: 23)), isA<TimeOnly>());
+      expect(TimeOnly.fromDuration(Duration.zero), isA<TimeOnly>());
+      expect(TimeOnly.fromDuration(const Duration(hours: 23)), isA<TimeOnly>());
     });
 
     test('rejects wrong durations', () {
-      expect(() => TimeOnly(const Duration(days: 1)), throwsAssertionError);
-      expect(() => TimeOnly(const Duration(days: 24)), throwsAssertionError);
-      expect(() => TimeOnly(-const Duration(days: 1)), throwsAssertionError);
       expect(
-        () => TimeOnly(-const Duration(milliseconds: 1)),
+        () => TimeOnly.fromDuration(const Duration(days: 1)),
+        throwsAssertionError,
+      );
+      expect(
+        () => TimeOnly.fromDuration(const Duration(days: 24)),
+        throwsAssertionError,
+      );
+      expect(
+        () => TimeOnly.fromDuration(-const Duration(days: 1)),
+        throwsAssertionError,
+      );
+      expect(
+        () => TimeOnly.fromDuration(-const Duration(milliseconds: 1)),
         throwsAssertionError,
       );
     });
@@ -32,7 +43,7 @@ void main() {
 
       expect(
         TimeOnly.fromDateTime(dt1),
-        TimeOnly(
+        TimeOnly.fromDuration(
           const Duration(
             hours: 6,
             minutes: 54,
@@ -44,7 +55,7 @@ void main() {
       );
       expect(
         TimeOnly.fromDateTime(dt2),
-        TimeOnly(
+        TimeOnly.fromDuration(
           const Duration(
             hours: 11,
             minutes: 54,
@@ -56,7 +67,7 @@ void main() {
       );
       expect(
         TimeOnly.fromDateTime(dt3),
-        TimeOnly(
+        TimeOnly.fromDuration(
           const Duration(
             hours: 8,
             minutes: 54,
@@ -69,8 +80,8 @@ void main() {
     });
 
     test('time part getters', () {
-      final zero = TimeOnly(Duration.zero);
-      final complex = TimeOnly(
+      final zero = TimeOnly.fromDuration(Duration.zero);
+      final complex = TimeOnly.fromDuration(
         const Duration(hours: 23, minutes: 32, seconds: 42, milliseconds: 1),
       );
 
@@ -85,15 +96,27 @@ void main() {
       expect(complex.microsecond, 1000);
     });
 
-    test('serialization', () {
-      expect(TimeOnly(Duration.zero).toJson(), '00:00:00.000000');
+    test('toDuration', () {
+      const zero = TimeOnly(0, 0, 0, 0);
+      const complex = TimeOnly(23, 32, 42, 1000);
+
+      expect(zero.toDuration(), Duration.zero);
       expect(
-        TimeOnly(const Duration(days: 1) - const Duration(microseconds: 1))
-            .toJson(),
+        complex.toDuration(),
+        const Duration(hours: 23, minutes: 32, seconds: 42, milliseconds: 1),
+      );
+    });
+
+    test('serialization', () {
+      expect(TimeOnly.fromDuration(Duration.zero).toJson(), '00:00:00.000000');
+      expect(
+        TimeOnly.fromDuration(
+          const Duration(days: 1) - const Duration(microseconds: 1),
+        ).toJson(),
         '23:59:59.999999',
       );
       expect(
-        TimeOnly(
+        TimeOnly.fromDuration(
           const Duration(hours: 23, minutes: 32, seconds: 42, milliseconds: 1),
         ).toJson(),
         '23:32:42.001000',
@@ -101,14 +124,19 @@ void main() {
     });
 
     test('deserialization', () {
-      expect(TimeOnly.fromJson('00:00:00.000000'), TimeOnly(Duration.zero));
+      expect(
+        TimeOnly.fromJson('00:00:00.000000'),
+        TimeOnly.fromDuration(Duration.zero),
+      );
       expect(
         TimeOnly.fromJson('23:59:59.999999'),
-        TimeOnly(const Duration(days: 1) - const Duration(microseconds: 1)),
+        TimeOnly.fromDuration(
+          const Duration(days: 1) - const Duration(microseconds: 1),
+        ),
       );
       expect(
         TimeOnly.fromJson('23:32:42.001000'),
-        TimeOnly(
+        TimeOnly.fromDuration(
           const Duration(hours: 23, minutes: 32, seconds: 42, milliseconds: 1),
         ),
       );
@@ -117,24 +145,24 @@ void main() {
     group('compareTo', () {
       test('greater than', () {
         expect(
-          TimeOnly(const Duration(hours: 1))
-              .compareTo(TimeOnly(const Duration(minutes: 1))),
+          TimeOnly.fromDuration(const Duration(hours: 1))
+              .compareTo(TimeOnly.fromDuration(const Duration(minutes: 1))),
           isPositive,
         );
       });
 
       test('less than', () {
         expect(
-          TimeOnly(const Duration(minutes: 1))
-              .compareTo(TimeOnly(const Duration(hours: 1))),
+          TimeOnly.fromDuration(const Duration(minutes: 1))
+              .compareTo(TimeOnly.fromDuration(const Duration(hours: 1))),
           isNegative,
         );
       });
 
       test('equal', () {
         expect(
-          TimeOnly(const Duration(hours: 1))
-              .compareTo(TimeOnly(const Duration(hours: 1))),
+          TimeOnly.fromDuration(const Duration(hours: 1))
+              .compareTo(TimeOnly.fromDuration(const Duration(hours: 1))),
           isZero,
         );
       });