+++ packages/leancode_contracts/CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.6.0
 
 - **BREAKING**: Removed `DateOnly.source`. Use `DateOnly.toDateTimeLocal()` or `DateOnly.toDateTimeUtc()` instead
+- **BREAKING**: Removed `TimeOnly.source`. Use `DateOnly.toDuration()` instead
+- **BREAKING**: `TimeOnly` constructor now accepts individual components instead of a `Duration` object. Use `TimeOnly.fromDuration` instead
 - Make `DateOnly` const-constructable
 
 # 0.5.0